kubernetes/website_with_db/db_read/requirements.txt
flask
google-cloud-firestore
pytest